backend/requirements.txt
Django==4.0.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.1.0
pymongo==4.3.3
djongo==1.3.6
drf-yasg==1.21.7
django-filter==23.2
python-decouple==3.8
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
